\encoding{UTF-8}
\name{banks76}
\alias{banks76}
\docType{data}
\title{
Data from Arthur Banks CNTS Data Archive for 1976}
\description{
Data from the Banks CNTS Data Archive for 1976.  These data are primarily used to demonstrate the \code{optMVtrans} function.  They have been listwise deleted and scaled according to the guidelines in the codebook.
}
\usage{data(banks76)}
\format{
  A data frame with 80 observations on the following 11 variables.
  \describe{
    \item{\code{ccode}}{Correlates of War Country Code}
    \item{\code{year}}{Year}
    \item{\code{country}}{Character Country Name}
    \item{\code{importspc}}{Imports/capita}
    \item{\code{exportspc}}{Exports/capita}
    \item{\code{enprodkgpc}}{Energy Production in kg/capita}
    \item{\code{enconskgpc}}{Energy Consumption in kg/capita}
    \item{\code{wfpcgind}}{Percentage of Work Force in Industry}
    \item{\code{newspaper}}{Newspaper Circulation}
    \item{\code{literate}}{Literacy Rate}
    \item{\code{gdppcmp}}{GDP/Capita (Market Prices)}
  }
}
\source{
http://www.databanksinternational.com
}
\references{
Banks, Arthur S. 2011. Cross-National Time-Series Data Archive. Databanks International. Jerusalem, Israel.
}
\keyword{datasets}
