\name{icpsrmsda-package}
\alias{icpsrmsda-package}
\alias{icpsrmsda}
\docType{package}
\title{
Scaling Measurement and Dimensional Analysis
}
\description{
This package implements functions designed for the Scaling, Measurement and Dimensional Analysis course at the ICPSR Summer Program for Quantitative Methods. 
}
\details{
\tabular{ll}{
Package: \tab icpsrmsda\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2013-07-31\cr
License: \tab GPL (>=2.0)\cr
LazyLoad: \tab yes\cr
}
This package is largely designed to support students in the Scaling, Measurement and Dimensional Analysis course.  Most functions for undertaking analysis in the course exist in other packages, but these functions comprise a set of utility and auxiliary functions. 
}
\author{
Dave Armstrong and Bill Jacoby\cr
Maintainer: Dave Armstrong <davearmstrong.ps@gmail.com>
}
\keyword{Multivariate Statistics}
